+++ maelstrom/grpc_provider.go
@@ -130,7 +130,7 @@ func (m *maelstromGrpcProvider) HandleClientRequest(msgType MsgType, msg any) {
 	case MsgTypeWrite:
 		w := msg.(*Message[Write])
 		if res, err := m.getService(oxiaClient).(proto.OxiaClientServer).Write(context.Background(), &proto.WriteRequest{
-			ShardId: pb.Uint32(0),
+			ShardId: pb.Int64(0),
 			Puts: []*proto.PutRequest{{
 				Key:               fmt.Sprintf("%d", w.Body.Key),
 				Value:             []byte(fmt.Sprintf("%d", w.Body.Value)),
@@ -158,7 +158,7 @@ func (m *maelstromGrpcProvider) HandleClientRequest(msgType MsgType, msg any) {
 		r := msg.(*Message[Read])
 		stream := newMaelstromReadServerStream()
 		err := m.getService(oxiaClient).(proto.OxiaClientServer).Read(&proto.ReadRequest{
-			ShardId: pb.Uint32(0),
+			ShardId: pb.Int64(0),
 			Gets: []*proto.GetRequest{{
 				Key:          fmt.Sprintf("%d", r.Body.Key),
 				IncludeValue: true,
@@ -196,7 +196,7 @@ func (m *maelstromGrpcProvider) HandleClientRequest(msgType MsgType, msg any) {
 		c := msg.(*Message[Cas])
 		stream := newMaelstromReadServerStream()
 		err := m.getService(oxiaClient).(proto.OxiaClientServer).Read(&proto.ReadRequest{
-			ShardId: pb.Uint32(0),
+			ShardId: pb.Int64(0),
 			Gets: []*proto.GetRequest{{
 				Key:          fmt.Sprintf("%d", c.Body.Key),
 				IncludeValue: true,
@@ -225,7 +225,7 @@ func (m *maelstromGrpcProvider) HandleClientRequest(msgType MsgType, msg any) {
 
 		// Write it back with conditional write
 		if writeRes, err := m.getService(oxiaClient).(proto.OxiaClientServer).Write(context.Background(), &proto.WriteRequest{
-			ShardId: pb.Uint32(0),
+			ShardId: pb.Int64(0),
 			Puts: []*proto.PutRequest{{
 				Key:               fmt.Sprintf("%d", c.Body.Key),
 				Value:             []byte(fmt.Sprintf("%d", c.Body.To)),
